+++ Explorer/Assets/DCL/PluginSystem/World/Dependencies/ECSWorldInstanceSharedDependencies.cs
@@ -2,6 +2,7 @@
 using CRDT;
 using CrdtEcsBridge.ECSToCRDTWriter;
 using DCL.Interaction.Utility;
+using DCL.Time;
 using ECS.Prioritization.Components;
 using SceneRunner.Scene;
 using SceneRunner.Scene.ExceptionsHandling;
@@ -20,6 +21,7 @@ namespace DCL.PluginSystem.World.Dependencies
         public readonly IEntityCollidersSceneCache EntityCollidersSceneCache;
         public readonly ISceneStateProvider SceneStateProvider;
         public readonly MutexSync MutexSync;
+        public readonly IWorldTimeProvider WorldTimeProvider;
 
         public ECSWorldInstanceSharedDependencies(
             ISceneData sceneData,
@@ -28,7 +30,8 @@ namespace DCL.PluginSystem.World.Dependencies
             Dictionary<CRDTEntity, Entity> entitiesMap,
             ISceneExceptionsHandler sceneExceptionsHandler,
             IEntityCollidersSceneCache entityCollidersSceneCache,
-            ISceneStateProvider sceneStateProvider, MutexSync mutexSync)
+            ISceneStateProvider sceneStateProvider, MutexSync mutexSync,
+            IWorldTimeProvider worldTimeProvider)
         {
             SceneData = sceneData;
             EcsToCRDTWriter = ecsToCRDTWriter;
@@ -39,6 +42,7 @@ namespace DCL.PluginSystem.World.Dependencies
             SceneExceptionsHandler = sceneExceptionsHandler;
             EntityCollidersSceneCache = entityCollidersSceneCache;
             SceneStateProvider = sceneStateProvider;
+            WorldTimeProvider = worldTimeProvider;
         }
     }
 }
